store photos of mowies

+++ df1/app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\Movie;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -92,7 +93,16 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->hasMovies()->count()) {
+            return 'NoNoNo';
+        }
         $category->delete();
+        return redirect()->route('c_index');
+    }
+    public function destroyAll(Category $category)
+    {
+        $ids = $category->hasMovies()->pluck('id')->all();
+        Movie::destroy($ids);
         return redirect()->route('c_index');
     }
 }
